+++ src/layouts/dashboard/SideBar.js
@@ -1,0 +1,136 @@
+import React, { useState } from "react";
+import { Avatar, Box, Divider, IconButton, Stack } from "@mui/material";
+import { useTheme } from "@mui/material/styles";
+import { Gear } from "phosphor-react";
+import AntSwitch from "../../components/AntSwitch";
+// import MaterialUISwitch from "../../components/MuiSwitch";
+import { Nav_Buttons } from "../../data";
+import useSettings from "../../hooks/useSettings";
+import { faker } from "@faker-js/faker";
+import Logo from "../../assets/Images/logo.ico";
+
+const SideBar = () => {
+  const theme = useTheme();
+  const [selected, setSelected] = useState(0);
+  const { onToggleMode } = useSettings();
+  return (
+    <Box
+      p={2}
+      sx={{
+        backgroundColor: theme.palette.background.paper,
+        height: "100vh",
+        width: 100,
+        boxShadow: "0px 0px 2px rgba(0,0,0,0.25)",
+      }}
+    >
+      <Stack
+        direction={"column"}
+        sx={{ height: "100%" }}
+        alignItems={"center"}
+        spacing={3}
+        justifyContent="space-between"
+      >
+        <Stack alignItems={"center"} spacing={4}>
+          {/* // for chat app logo */}
+          <Box
+            sx={{
+              backgroundColor: theme.palette.primary.main,
+              height: 64,
+              width: 64,
+              borderRadius: 1.5,
+            }}
+          >
+            <img src={Logo} alt="chat app logo" />
+          </Box>
+          {/* // for Nav_Buttons */}
+          <Stack
+            spacing={3}
+            sx={{ width: "max-content" }}
+            direction="column"
+            alignItems="center"
+          >
+            {Nav_Buttons.map((el) =>
+              // if el.index === selected, then render with box to showcase background colored(blue) as selected
+              el.index === selected ? (
+                <Box
+                  sx={{
+                    backgroundColor: theme.palette.primary.main,
+                    borderRadius: 1.5,
+                  }}
+                >
+                  <IconButton
+                    sx={{ width: "max-content", color: "#fff" }}
+                    key={el.index}
+                  >
+                    {el.icon}
+                  </IconButton>
+                </Box>
+              ) : (
+                <IconButton
+                  onClick={() => {
+                    setSelected(el.index);
+                  }}
+                  sx={{
+                    width: "max-content",
+                    color:
+                      theme.palette.mode === "light"
+                        ? "#000"
+                        : theme.palette.text.primary,
+                  }}
+                  key={el.index}
+                >
+                  {el.icon}
+                </IconButton>
+              )
+            )}
+            <Divider sx={{ width: "48px" }} />
+            {selected === 3 ? (
+              <Box
+                sx={{
+                  backgroundColor: theme.palette.primary.main,
+                  borderRadius: 1.5,
+                }}
+              >
+                <IconButton sx={{ width: "max-content", color: "#fff" }}>
+                  <Gear />
+                </IconButton>
+              </Box>
+            ) : (
+              <IconButton
+                onClick={() => {
+                  setSelected(3);
+                }}
+                sx={{
+                  width: "max-content",
+                  color:
+                    theme.palette.mode === "light"
+                      ? "#000"
+                      : theme.palette.text.primary,
+                }}
+              >
+                <Gear />
+              </IconButton>
+            )}
+          </Stack>
+        </Stack>
+        <Stack alignItems={"center"} spacing={3}>
+          {/* <MaterialUISwitch
+            onChange={() => {
+              onToggleMode();
+            }}
+            defaultChecked
+          /> */}
+          <AntSwitch
+            onChange={() => {
+              onToggleMode();
+            }}
+            defaultChecked
+          />
+          <Avatar src={faker.image.avatar()} />
+        </Stack>
+      </Stack>
+    </Box>
+  );
+};
+
+export default SideBar;
